Move balanceTx to new `Cardano.Wallet.Write.Tx.Balance`
- Drop ErrNoSuchWallet from ErrSelectAssets to avoid cyclic imports
- Split off BalanceTxLog
- Slight changes to import names etc
